requirements.txt
chromadb
openai
anthropic
google-generativeai
ultralytics
opencv-python
pandas
regex
pillow
numpy
argparse